requirements.txt
Flask==1.1.2
Flask_JSON==0.3.4
jsonschema==3.2.0
marshmallow==3.9.1
Flask-Cors==3.0.9
Flask-SocketIO==4.3.1
Flask-SQLAlchemy==2.4.4
Flask-Migrate==2.5.3
eventlet==0.29.1
psycopg2-binary==2.8.6
Flask-Session==0.3.2
